// src/lib2/models/File.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type File = {
    columns?: Array<string>;
    columns_count?: number;
    created_at?: string;
    id?: string;
    mime_type?: string;
    name?: string;
    rows_count?: number;
    /**
     * Size in bytes
     */
    size?: number;
    workspace_id?: string;
};